ref: improve auth code sharing
